% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{approve_skill}
\alias{approve_skill}
\title{Associates a skill with the organization under the customer's AWS account}
\usage{
approve_skill(
  SkillId = "string"
)
}
\arguments{
\item{SkillId}{[required] The unique identifier of the skill.}
}
\description{
Associates a skill with the organization under the customer's AWS account. If a skill is private, the user implicitly accepts access to this skill during enablement.
}
\examples{

}
