% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{associate_device_with_room}
\alias{associate_device_with_room}
\title{Associates a device with a given room}
\usage{
associate_device_with_room(
  DeviceArn = "string",
  RoomArn = "string"
)
}
\arguments{
\item{DeviceArn}{The ARN of the device to associate to a room. Required.}

\item{RoomArn}{The ARN of the room with which to associate the device. Required.}
}
\description{
Associates a device with a given room. This applies all the settings from the room profile to the device, and all the skills in any skill groups added to that room. This operation requires the device to be online, or else a manual sync is required.
}
\examples{

}
