% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{associate_skill_group_with_room}
\alias{associate_skill_group_with_room}
\title{Associates a skill group with a given room}
\usage{
associate_skill_group_with_room(
  SkillGroupArn = "string",
  RoomArn = "string"
)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to associate with a room. Required.}

\item{RoomArn}{The ARN of the room with which to associate the skill group. Required.}
}
\description{
Associates a skill group with a given room. This enables all skills in the associated skill group on all devices in the room.
}
\examples{

}
