% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{associate_skill_with_skill_group}
\alias{associate_skill_with_skill_group}
\title{Associates a skill with a skill group}
\usage{
associate_skill_with_skill_group(SkillGroupArn = NULL, SkillId)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to associate the skill to. Required.}

\item{SkillId}{[required] The unique identifier of the skill.}
}
\description{
Associates a skill with a skill group.
}
\section{Accepted Parameters}{
\preformatted{associate_skill_with_skill_group(
  SkillGroupArn = "string",
  SkillId = "string"
)
}
}

