% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{associate_skill_with_users}
\alias{associate_skill_with_users}
\title{Makes a private skill available for enrolled users to enable on their devices}
\usage{
associate_skill_with_users(OrganizationArn = NULL, SkillId)
}
\arguments{
\item{OrganizationArn}{The ARN of the organization.}

\item{SkillId}{[required] The private skill ID you want to make available to enrolled users.>}
}
\description{
Makes a private skill available for enrolled users to enable on their devices.
}
\section{Accepted Parameters}{
\preformatted{associate_skill_with_users(
  OrganizationArn = "string",
  SkillId = "string"
)
}
}

