% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_address_book}
\alias{create_address_book}
\title{Creates an address book with the specified details}
\usage{
create_address_book(Name, Description = NULL, ClientRequestToken = NULL)
}
\arguments{
\item{Name}{The name of the address book.}

\item{Description}{The description of the address book.}

\item{ClientRequestToken}{A unique, user-specified identifier for the request that ensures idempotency.}
}
\description{
Creates an address book with the specified details.
}
\examples{

}
