% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_contact}
\alias{create_contact}
\title{Creates a contact with the specified details}
\usage{
create_contact(DisplayName = NULL, FirstName, LastName = NULL,
  PhoneNumber = NULL, ClientRequestToken = NULL)
}
\arguments{
\item{DisplayName}{The name of the contact to display on the console.}

\item{FirstName}{[required] The first name of the contact that is used to call the contact on the device.}

\item{LastName}{The last name of the contact that is used to call the contact on the device.}

\item{PhoneNumber}{The phone number of the contact in E.164 format.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures idempotency.}
}
\description{
Creates a contact with the specified details.
}
\section{Accepted Parameters}{
\preformatted{create_contact(
  DisplayName = "string",
  FirstName = "string",
  LastName = "string",
  PhoneNumber = "string",
  ClientRequestToken = "string"
)
}
}

