% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_profile}
\alias{create_profile}
\title{Creates a new room profile with the specified details}
\usage{
create_profile(ProfileName, Timezone, Address, DistanceUnit, TemperatureUnit,
  WakeWord, ClientRequestToken = NULL, SetupModeDisabled = NULL,
  MaxVolumeLimit = NULL, PSTNEnabled = NULL)
}
\arguments{
\item{ProfileName}{The name of a room profile.}

\item{Timezone}{The time zone used by a room profile.}

\item{Address}{The valid address for the room.}

\item{DistanceUnit}{The distance unit to be used by devices in the profile.}

\item{TemperatureUnit}{The temperature unit to be used by devices in the profile.}

\item{WakeWord}{A wake word for Alexa, Echo, Amazon, or a computer.}

\item{ClientRequestToken}{The user-specified token that is used during the creation of a profile.}

\item{SetupModeDisabled}{Whether room profile setup is enabled.}

\item{MaxVolumeLimit}{The maximum volume limit for a room profile.}

\item{PSTNEnabled}{Whether PSTN calling is enabled.}
}
\description{
Creates a new room profile with the specified details.
}
\examples{

}
