% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_skill_group}
\alias{create_skill_group}
\title{Creates a skill group with a specified name and description}
\usage{
create_skill_group(SkillGroupName, Description = NULL,
  ClientRequestToken = NULL)
}
\arguments{
\item{SkillGroupName}{The name for the skill group.}

\item{Description}{The description for the skill group.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures idempotency.}
}
\description{
Creates a skill group with a specified name and description.
}
\examples{

}
