% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a user}
\usage{
create_user(UserId, FirstName = NULL, LastName = NULL, Email = NULL,
  ClientRequestToken = NULL, Tags = NULL)
}
\arguments{
\item{UserId}{[required] The ARN for the user.}

\item{FirstName}{The first name for the user.}

\item{LastName}{The last name for the user.}

\item{Email}{The email address for the user.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures idempotency.}

\item{Tags}{The tags for the user.}
}
\description{
Creates a user.
}
\section{Accepted Parameters}{
\preformatted{create_user(
  UserId = "string",
  FirstName = "string",
  LastName = "string",
  Email = "string",
  ClientRequestToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

