% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_conference_provider}
\alias{delete_conference_provider}
\title{Deletes a conference provider}
\usage{
delete_conference_provider(ConferenceProviderArn)
}
\arguments{
\item{ConferenceProviderArn}{[required] The ARN of the conference provider.}
}
\description{
Deletes a conference provider.
}
\section{Accepted Parameters}{
\preformatted{delete_conference_provider(
  ConferenceProviderArn = "string"
)
}
}

