% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_contact}
\alias{delete_contact}
\title{Deletes a contact by the contact ARN}
\usage{
delete_contact(ContactArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to delete.}
}
\description{
Deletes a contact by the contact ARN.
}
\section{Accepted Parameters}{
\preformatted{delete_contact(
  ContactArn = "string"
)
}
}

