% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_room}
\alias{delete_room}
\title{Deletes a room by the room ARN}
\usage{
delete_room(RoomArn = NULL)
}
\arguments{
\item{RoomArn}{The ARN of the room to delete. Required.}
}
\description{
Deletes a room by the room ARN.
}
\examples{

}
