% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_skill_authorization}
\alias{delete_skill_authorization}
\title{Unlinks a third-party account from a skill}
\usage{
delete_skill_authorization(
  SkillId = "string",
  RoomArn = "string"
)
}
\arguments{
\item{SkillId}{[required] The unique identifier of a skill.}

\item{RoomArn}{The room that the skill is authorized for.}
}
\description{
Unlinks a third-party account from a skill.
}
\examples{

}
