% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_skill_group}
\alias{delete_skill_group}
\title{Deletes a skill group by skill group ARN}
\usage{
delete_skill_group(SkillGroupArn = NULL)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to delete. Required.}
}
\description{
Deletes a skill group by skill group ARN.
}
\section{Accepted Parameters}{
\preformatted{delete_skill_group(
  SkillGroupArn = "string"
)
}
}

