% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_user}
\alias{delete_user}
\title{Deletes a specified user by user ARN and enrollment ARN}
\usage{
delete_user(
  UserArn = "string",
  EnrollmentId = "string"
)
}
\arguments{
\item{UserArn}{The ARN of the user to delete in the organization. Required.}

\item{EnrollmentId}{[required] The ARN of the user's enrollment in the organization. Required.}
}
\description{
Deletes a specified user by user ARN and enrollment ARN.
}
\examples{

}
