% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{disassociate_device_from_room}
\alias{disassociate_device_from_room}
\title{Disassociates a device from its current room}
\usage{
disassociate_device_from_room(DeviceArn = NULL)
}
\arguments{
\item{DeviceArn}{The ARN of the device to disassociate from a room. Required.}
}
\description{
Disassociates a device from its current room. The device continues to be connected to the Wi-Fi network and is still registered to the account. The device settings and skills are removed from the room.
}
\examples{

}
