% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{disassociate_skill_from_skill_group}
\alias{disassociate_skill_from_skill_group}
\title{Disassociates a skill from a skill group}
\usage{
disassociate_skill_from_skill_group(SkillGroupArn = NULL, SkillId)
}
\arguments{
\item{SkillGroupArn}{The unique identifier of a skill. Required.}

\item{SkillId}{[required] The ARN of a skill group to associate to a skill.}
}
\description{
Disassociates a skill from a skill group.
}
\section{Accepted Parameters}{
\preformatted{disassociate_skill_from_skill_group(
  SkillGroupArn = "string",
  SkillId = "string"
)
}
}

