% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{disassociate_skill_from_users}
\alias{disassociate_skill_from_users}
\title{Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices}
\usage{
disassociate_skill_from_users(OrganizationArn = NULL, SkillId)
}
\arguments{
\item{OrganizationArn}{The ARN of the organization.}

\item{SkillId}{[required] The private skill ID you want to make unavailable for enrolled users.}
}
\description{
Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices.
}
\section{Accepted Parameters}{
\preformatted{disassociate_skill_from_users(
  OrganizationArn = "string",
  SkillId = "string"
)
}
}

