% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{forget_smart_home_appliances}
\alias{forget_smart_home_appliances}
\title{Forgets smart home appliances associated to a room}
\usage{
forget_smart_home_appliances(RoomArn)
}
\arguments{
\item{RoomArn}{[required] The room that the appliances are associated with.}
}
\description{
Forgets smart home appliances associated to a room.
}
\section{Accepted Parameters}{
\preformatted{forget_smart_home_appliances(
  RoomArn = "string"
)
}
}

