% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_address_book}
\alias{get_address_book}
\title{Gets address the book details by the address book ARN}
\usage{
get_address_book(
  AddressBookArn = "string"
)
}
\arguments{
\item{AddressBookArn}{[required] The ARN of the address book for which to request details.}
}
\description{
Gets address the book details by the address book ARN.
}
\examples{

}
