% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_conference_preference}
\alias{get_conference_preference}
\title{Retrieves the existing conference preferences}
\usage{
get_conference_preference()
}
\description{
Retrieves the existing conference preferences.
}
\examples{

}
