% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_conference_provider}
\alias{get_conference_provider}
\title{Gets details about a specific conference provider}
\usage{
get_conference_provider(ConferenceProviderArn)
}
\arguments{
\item{ConferenceProviderArn}{[required] The ARN of the newly created conference provider.}
}
\description{
Gets details about a specific conference provider.
}
\section{Accepted Parameters}{
\preformatted{get_conference_provider(
  ConferenceProviderArn = "string"
)
}
}

