% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_contact}
\alias{get_contact}
\title{Gets the contact details by the contact ARN}
\usage{
get_contact(ContactArn)
}
\arguments{
\item{ContactArn}{The ARN of the contact for which to request details.}
}
\description{
Gets the contact details by the contact ARN.
}
\examples{

}
