% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_device}
\alias{get_device}
\title{Gets the details of a device by device ARN}
\usage{
get_device(DeviceArn = NULL)
}
\arguments{
\item{DeviceArn}{The ARN of the device for which to request details. Required.}
}
\description{
Gets the details of a device by device ARN.
}
\examples{

}
