% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_profile}
\alias{get_profile}
\title{Gets the details of a room profile by profile ARN}
\usage{
get_profile(ProfileArn = NULL)
}
\arguments{
\item{ProfileArn}{The ARN of the room profile for which to request details. Required.}
}
\description{
Gets the details of a room profile by profile ARN.
}
\examples{

}
