% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_room}
\alias{get_room}
\title{Gets room details by room ARN}
\usage{
get_room(RoomArn = NULL)
}
\arguments{
\item{RoomArn}{The ARN of the room for which to request details. Required.}
}
\description{
Gets room details by room ARN.
}
\section{Accepted Parameters}{
\preformatted{get_room(
  RoomArn = "string"
)
}
}

