% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_room_skill_parameter}
\alias{get_room_skill_parameter}
\title{Gets room skill parameter details by room, skill, and parameter key ARN}
\usage{
get_room_skill_parameter(SkillId, ParameterKey, RoomArn = NULL)
}
\arguments{
\item{SkillId}{The ARN of the skill from which to get the room skill parameter details. Required.}

\item{ParameterKey}{The room skill parameter key for which to get details. Required.}

\item{RoomArn}{The ARN of the room from which to get the room skill parameter details.}
}
\description{
Gets room skill parameter details by room, skill, and parameter key ARN.
}
\examples{

}
