% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{get_skill_group}
\alias{get_skill_group}
\title{Gets skill group details by skill group ARN}
\usage{
get_skill_group(SkillGroupArn = NULL)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group for which to get details. Required.}
}
\description{
Gets skill group details by skill group ARN.
}
\section{Accepted Parameters}{
\preformatted{get_skill_group(
  SkillGroupArn = "string"
)
}
}

