% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_conference_providers}
\alias{list_conference_providers}
\title{Lists conference providers under a specific AWS account}
\usage{
list_conference_providers(
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of conference providers to be returned, per paginated calls.}
}
\description{
Lists conference providers under a specific AWS account.
}
\examples{

}
