% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_device_events}
\alias{list_device_events}
\title{Lists the device event history, including device connection status, for up to 30 days}
\usage{
list_device_events(DeviceArn, EventType = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{DeviceArn}{[required] The ARN of a device.}

\item{EventType}{The event type to filter device events. If EventType isn't specified, this returns a list of all device events in reverse chronological order. If EventType is specified, this returns a list of device events for that EventType in reverse chronological order.}

\item{NextToken}{An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults. When the end of results is reached, the response has a value of null.}

\item{MaxResults}{The maximum number of results to include in the response. The default value is 50. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.}
}
\description{
Lists the device event history, including device connection status, for up to 30 days.
}
\section{Accepted Parameters}{
\preformatted{list_device_events(
  DeviceArn = "string",
  EventType = "CONNECTION_STATUS"|"DEVICE_STATUS",
  NextToken = "string",
  MaxResults = 123
)
}
}

