% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_skills_store_categories}
\alias{list_skills_store_categories}
\title{Lists all categories in the Alexa skill store}
\usage{
list_skills_store_categories(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of categories returned, per paginated calls.}
}
\description{
Lists all categories in the Alexa skill store.
}
\section{Accepted Parameters}{
\preformatted{list_skills_store_categories(
  NextToken = "string",
  MaxResults = 123
)
}
}

