% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_skills_store_skills_by_category}
\alias{list_skills_store_skills_by_category}
\title{Lists all skills in the Alexa skill store by category}
\usage{
list_skills_store_skills_by_category(CategoryId, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{CategoryId}{[required] The category ID for which the skills are being retrieved from the skill store.}

\item{NextToken}{The tokens used for pagination.}

\item{MaxResults}{The maximum number of skills returned per paginated calls.}
}
\description{
Lists all skills in the Alexa skill store by category.
}
\section{Accepted Parameters}{
\preformatted{list_skills_store_skills_by_category(
  CategoryId = 123,
  NextToken = "string",
  MaxResults = 123
)
}
}

