% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_smart_home_appliances}
\alias{list_smart_home_appliances}
\title{Lists all of the smart home appliances associated with a room}
\usage{
list_smart_home_appliances(RoomArn, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{RoomArn}{[required] The room that the appliances are associated with.}

\item{MaxResults}{The maximum number of appliances to be returned, per paginated calls.}

\item{NextToken}{The tokens used for pagination.}
}
\description{
Lists all of the smart home appliances associated with a room.
}
\section{Accepted Parameters}{
\preformatted{list_smart_home_appliances(
  RoomArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

