% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{list_tags}
\alias{list_tags}
\title{Lists all tags for the specified resource}
\usage{
list_tags(
  Arn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{Arn}{[required] The ARN of the specified resource for which to list tags.}

\item{NextToken}{An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by \code{MaxResults}.}

\item{MaxResults}{The maximum number of results to include in the response. If more results exist than the specified \code{MaxResults} value, a token is included in the response so that the remaining results can be retrieved.}
}
\description{
Lists all tags for the specified resource.
}
\examples{

}
