% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{put_conference_preference}
\alias{put_conference_preference}
\title{Sets the conference preferences on a specific conference provider at the account level}
\usage{
put_conference_preference(ConferencePreference)
}
\arguments{
\item{ConferencePreference}{[required] The conference preference of a specific conference provider.}
}
\description{
Sets the conference preferences on a specific conference provider at the account level.
}
\section{Accepted Parameters}{
\preformatted{put_conference_preference(
  ConferencePreference = list(
    DefaultConferenceProviderArn = "string"
  )
)
}
}

