% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{put_skill_authorization}
\alias{put_skill_authorization}
\title{Links a user's account to a third-party skill provider}
\usage{
put_skill_authorization(AuthorizationResult, SkillId, RoomArn = NULL)
}
\arguments{
\item{AuthorizationResult}{[required] The authorization result specific to OAUTH code grant output. "Code" must be populated in the AuthorizationResult map to establish the authorization.}

\item{SkillId}{[required] The unique identifier of a skill.}

\item{RoomArn}{The room that the skill is authorized for.}
}
\description{
Links a user's account to a third-party skill provider. If this API operation is called by an assumed IAM role, the skill being linked must be a private skill. Also, the skill must be owned by the AWS account that assumed the IAM role.
}
\section{Accepted Parameters}{
\preformatted{put_skill_authorization(
  AuthorizationResult = list(
    "string"
  ),
  SkillId = "string",
  RoomArn = "string"
)
}
}

