% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{register_avs_device}
\alias{register_avs_device}
\title{Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS)}
\usage{
register_avs_device(ClientId, UserCode, ProductId, DeviceSerialNumber,
  AmazonId)
}
\arguments{
\item{ClientId}{[required] The client ID of the OEM used for code-based linking authorization on an AVS device.}

\item{UserCode}{[required] The code that is obtained after your AVS device has made a POST request to LWA as a part of the Device Authorization Request component of the OAuth code-based linking specification.}

\item{ProductId}{[required] The product ID used to identify your AVS device during authorization.}

\item{DeviceSerialNumber}{[required] The key generated by the OEM that uniquely identifies a specified instance of your AVS device.}

\item{AmazonId}{[required] The device type ID for your AVS device generated by Amazon when the OEM creates a new product on Amazon's Developer Console.}
}
\description{
Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS).
}
\section{Accepted Parameters}{
\preformatted{register_avs_device(
  ClientId = "string",
  UserCode = "string",
  ProductId = "string",
  DeviceSerialNumber = "string",
  AmazonId = "string"
)
}
}

