% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{reject_skill}
\alias{reject_skill}
\title{Disassociates a skill from the organization under a user's AWS account}
\usage{
reject_skill(SkillId)
}
\arguments{
\item{SkillId}{[required] The unique identifier of the skill.}
}
\description{
Disassociates a skill from the organization under a user's AWS account. If the skill is a private skill, it moves to an AcceptStatus of PENDING. Any private or public skill that is rejected can be added later by calling the ApproveSkill API.
}
\section{Accepted Parameters}{
\preformatted{reject_skill(
  SkillId = "string"
)
}
}

