% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{resolve_room}
\alias{resolve_room}
\title{Determines the details for the room from which a skill request was invoked}
\usage{
resolve_room(UserId, SkillId)
}
\arguments{
\item{UserId}{The ARN of the user. Required.}

\item{SkillId}{The ARN of the skill that was requested. Required.}
}
\description{
Determines the details for the room from which a skill request was invoked. This operation is used by skill developers.
}
\examples{

}
