% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{revoke_invitation}
\alias{revoke_invitation}
\title{Revokes an invitation and invalidates the enrollment URL}
\usage{
revoke_invitation(UserArn = NULL, EnrollmentId = NULL)
}
\arguments{
\item{UserArn}{The ARN of the user for whom to revoke an enrollment invitation. Required.}

\item{EnrollmentId}{The ARN of the enrollment invitation to revoke. Required.}
}
\description{
Revokes an invitation and invalidates the enrollment URL.
}
\section{Accepted Parameters}{
\preformatted{revoke_invitation(
  UserArn = "string",
  EnrollmentId = "string"
)
}
}

