% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{start_device_sync}
\alias{start_device_sync}
\title{Resets a device and its account to the known default settings, by clearing all information and settings set by previous users}
\usage{
start_device_sync(
  RoomArn = "string",
  DeviceArn = "string",
  Features = list(
    "BLUETOOTH"|"VOLUME"|"NOTIFICATIONS"|"LISTS"|"SKILLS"|"ALL"
  )
)
}
\arguments{
\item{RoomArn}{The ARN of the room with which the device to sync is associated. Required.}

\item{DeviceArn}{The ARN of the device to sync. Required.}

\item{Features}{[required] Request structure to start the device sync. Required.}
}
\description{
Resets a device and its account to the known default settings, by clearing all information and settings set by previous users.
}
\examples{

}
