% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{start_smart_home_appliance_discovery}
\alias{start_smart_home_appliance_discovery}
\title{Initiates the discovery of any smart home appliances associated with the room}
\usage{
start_smart_home_appliance_discovery(
  RoomArn = "string"
)
}
\arguments{
\item{RoomArn}{[required] The room where smart home appliance discovery was initiated.}
}
\description{
Initiates the discovery of any smart home appliances associated with the room.
}
\examples{

}
