% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds metadata tags to a specified resource}
\usage{
tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{The ARN of the resource to which to add metadata tags. Required.}

\item{Tags}{The tags to be added to the specified resource. Do not provide system tags. Required.}
}
\description{
Adds metadata tags to a specified resource.
}
\examples{

}
