% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes metadata tags from a specified resource}
\usage{
untag_resource(Arn, TagKeys)
}
\arguments{
\item{Arn}{[required] The ARN of the resource from which to remove metadata tags. Required.}

\item{TagKeys}{[required] The tags to be removed from the specified resource. Do not provide system tags. Required.}
}
\description{
Removes metadata tags from a specified resource.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  Arn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

