% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_address_book}
\alias{update_address_book}
\title{Updates address book details by the address book ARN}
\usage{
update_address_book(
  AddressBookArn = "string",
  Name = "string",
  Description = "string"
)
}
\arguments{
\item{AddressBookArn}{[required] The ARN of the room to update.}

\item{Name}{The updated name of the room.}

\item{Description}{The updated description of the room.}
}
\description{
Updates address book details by the address book ARN.
}
\examples{

}
