% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_contact}
\alias{update_contact}
\title{Updates the contact details by the contact ARN}
\usage{
update_contact(ContactArn, DisplayName = NULL, FirstName = NULL,
  LastName = NULL, PhoneNumber = NULL)
}
\arguments{
\item{ContactArn}{The ARN of the contact to update.}

\item{DisplayName}{The updated display name of the contact.}

\item{FirstName}{The updated first name of the contact.}

\item{LastName}{The updated last name of the contact.}

\item{PhoneNumber}{The updated phone number of the contact.}
}
\description{
Updates the contact details by the contact ARN.
}
\examples{

}
