% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_device}
\alias{update_device}
\title{Updates the device name by device ARN}
\usage{
update_device(
  DeviceArn = "string",
  DeviceName = "string"
)
}
\arguments{
\item{DeviceArn}{The ARN of the device to update. Required.}

\item{DeviceName}{The updated device name. Required.}
}
\description{
Updates the device name by device ARN.
}
\examples{

}
