% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_room}
\alias{update_room}
\title{Updates room details by room ARN}
\usage{
update_room(RoomArn = NULL, RoomName = NULL, Description = NULL,
  ProviderCalendarId = NULL, ProfileArn = NULL)
}
\arguments{
\item{RoomArn}{The ARN of the room to update.}

\item{RoomName}{The updated name for the room.}

\item{Description}{The updated description for the room.}

\item{ProviderCalendarId}{The updated provider calendar ARN for the room.}

\item{ProfileArn}{The updated profile ARN for the room.}
}
\description{
Updates room details by room ARN.
}
\examples{

}
