% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_skill_group}
\alias{update_skill_group}
\title{Updates skill group details by skill group ARN}
\usage{
update_skill_group(
  SkillGroupArn = "string",
  SkillGroupName = "string",
  Description = "string"
)
}
\arguments{
\item{SkillGroupArn}{The ARN of the skill group to update.}

\item{SkillGroupName}{The updated name for the skill group.}

\item{Description}{The updated description for the skill group.}
}
\description{
Updates skill group details by skill group ARN.
}
\examples{

}
