% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{associate_contact_with_address_book}
\alias{associate_contact_with_address_book}
\title{Associates a contact with a given address book}
\usage{
associate_contact_with_address_book(ContactArn, AddressBookArn)
}
\arguments{
\item{ContactArn}{The ARN of the contact to associate with an address book.}

\item{AddressBookArn}{The ARN of the address book with which to associate the contact.}
}
\description{
Associates a contact with a given address book.
}
\examples{

}
