% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_contact}
\alias{create_contact}
\title{Creates a contact with the specified details}
\usage{
create_contact(FirstName, PhoneNumber, DisplayName = NULL, LastName = NULL,
  ClientRequestToken = NULL)
}
\arguments{
\item{FirstName}{The first name of the contact that is used to call the contact on the device.}

\item{PhoneNumber}{The phone number of the contact in E.164 format.}

\item{DisplayName}{The name of the contact to display on the console.}

\item{LastName}{The last name of the contact that is used to call the contact on the device.}

\item{ClientRequestToken}{A unique, user-specified identifier for this request that ensures idempotency.}
}
\description{
Creates a contact with the specified details.
}
\examples{

}
