% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{create_profile}
\alias{create_profile}
\title{Creates a new room profile with the specified details}
\usage{
create_profile(ProfileName, Timezone, Address, DistanceUnit,
  TemperatureUnit, WakeWord, ClientRequestToken = NULL,
  SetupModeDisabled = NULL, MaxVolumeLimit = NULL,
  PSTNEnabled = NULL)
}
\arguments{
\item{ProfileName}{[required] The name of a room profile.}

\item{Timezone}{[required] The time zone used by a room profile.}

\item{Address}{[required] The valid address for the room.}

\item{DistanceUnit}{[required] The distance unit to be used by devices in the profile.}

\item{TemperatureUnit}{[required] The temperature unit to be used by devices in the profile.}

\item{WakeWord}{[required] A wake word for Alexa, Echo, Amazon, or a computer.}

\item{ClientRequestToken}{The user-specified token that is used during the creation of a profile.}

\item{SetupModeDisabled}{Whether room profile setup is enabled.}

\item{MaxVolumeLimit}{The maximum volume limit for a room profile.}

\item{PSTNEnabled}{Whether PSTN calling is enabled.}
}
\description{
Creates a new room profile with the specified details.
}
\section{Accepted Parameters}{
\preformatted{create_profile(
  ProfileName = "string",
  Timezone = "string",
  Address = "string",
  DistanceUnit = "METRIC"|"IMPERIAL",
  TemperatureUnit = "FAHRENHEIT"|"CELSIUS",
  WakeWord = "ALEXA"|"AMAZON"|"ECHO"|"COMPUTER",
  ClientRequestToken = "string",
  SetupModeDisabled = TRUE|FALSE,
  MaxVolumeLimit = 123,
  PSTNEnabled = TRUE|FALSE
)
}
}

