% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_address_book}
\alias{delete_address_book}
\title{Deletes an address book by the address book ARN}
\usage{
delete_address_book(
  AddressBookArn = "string"
)
}
\arguments{
\item{AddressBookArn}{[required] The ARN of the address book to delete.}
}
\description{
Deletes an address book by the address book ARN.
}
\examples{

}
