% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_business_report_schedule}
\alias{delete_business_report_schedule}
\title{Deletes the recurring report delivery schedule with the specified schedule ARN}
\usage{
delete_business_report_schedule(ScheduleArn)
}
\arguments{
\item{ScheduleArn}{[required] The ARN of the business report schedule.}
}
\description{
Deletes the recurring report delivery schedule with the specified schedule ARN.
}
\section{Accepted Parameters}{
\preformatted{delete_business_report_schedule(
  ScheduleArn = "string"
)
}
}

