% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{delete_device}
\alias{delete_device}
\title{Removes a device from Alexa For Business}
\usage{
delete_device(
  DeviceArn = "string"
)
}
\arguments{
\item{DeviceArn}{[required] The ARN of the device for which to request details.}
}
\description{
Removes a device from Alexa For Business.
}
\examples{

}
