% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_package.R
\docType{package}
\name{paws.alexaforbusiness-package}
\alias{paws.alexaforbusiness}
\alias{paws.alexaforbusiness-package}
\title{paws.alexaforbusiness: Alexa For Business}
\description{
Alexa for Business helps you use Alexa in your organization.
Alexa for Business provides you with the tools to manage Alexa
devices, enroll your users, and assign skills, at scale. You can build
your own context-aware voice skills using the Alexa Skills Kit and the
Alexa for Business API operations. You can also make these available
as private skills for your organization. Alexa for Business makes it
efficient to voice-enable your products and services, thus providing
context-aware voice experiences for your customers. Device makers
building with the Alexa Voice Service (AVS) can create fully
integrated solutions, register their products with Alexa for Business,
and manage them as shared devices in their organization.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
