% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_package.R
\docType{package}
\name{paws.alexaforbusiness-package}
\alias{paws.alexaforbusiness}
\alias{paws.alexaforbusiness-package}
\title{paws.alexaforbusiness: Alexa For Business}
\description{
Alexa for Business makes it easy for you to use
Alexa in your organization. Alexa for Business gives you the tools you
need for managing Alexa devices, enroll your users, and assign skills,
at scale. You can build your own context-aware voice skills using the
Alexa Skills Kit and the Alexa for Business API operations. You can
make also these available as private skills for your organization.
Alexa for Business makes it easy to voice-enable your products and
services, providing context-aware voice experiences for your
customers.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
