% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.alexaforbusiness_operations.R
\name{update_profile}
\alias{update_profile}
\title{Updates an existing room profile by room profile ARN}
\usage{
update_profile(
  ProfileArn = "string",
  ProfileName = "string",
  IsDefault = TRUE|FALSE,
  Timezone = "string",
  Address = "string",
  DistanceUnit = "METRIC"|"IMPERIAL",
  TemperatureUnit = "FAHRENHEIT"|"CELSIUS",
  WakeWord = "ALEXA"|"AMAZON"|"ECHO"|"COMPUTER",
  SetupModeDisabled = TRUE|FALSE,
  MaxVolumeLimit = 123,
  PSTNEnabled = TRUE|FALSE
)
}
\arguments{
\item{ProfileArn}{The ARN of the room profile to update. Required.}

\item{ProfileName}{The updated name for the room profile.}

\item{IsDefault}{Sets the profile as default if selected. If this is missing, no update is done to the default status.}

\item{Timezone}{The updated timezone for the room profile.}

\item{Address}{The updated address for the room profile.}

\item{DistanceUnit}{The updated distance unit for the room profile.}

\item{TemperatureUnit}{The updated temperature unit for the room profile.}

\item{WakeWord}{The updated wake word for the room profile.}

\item{SetupModeDisabled}{Whether the setup mode of the profile is enabled.}

\item{MaxVolumeLimit}{The updated maximum volume limit for the room profile.}

\item{PSTNEnabled}{Whether the PSTN setting of the room profile is enabled.}
}
\description{
Updates an existing room profile by room profile ARN.
}
\examples{

}
